11usernamevalidation

+++ routes/ideas.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 const Idea = require('../models/Idea');
 
-//Get All Ideas
+//get all ideas
 router.get('/', async (req, res) => {
   try {
     const ideas = await Idea.find();
@@ -13,7 +13,7 @@ router.get('/', async (req, res) => {
   }
 });
 
-//Get a single idea
+//get a single idea
 router.get('/:id', async (req, res) => {
   try {
     const idea = await Idea.findById(req.params.id);
@@ -24,7 +24,7 @@ router.get('/:id', async (req, res) => {
   }
 });
 
-//Add an Idea
+// Add an idea
 router.post('/', async (req, res) => {
   const idea = new Idea({
     text: req.body.text,
@@ -41,31 +41,52 @@ router.post('/', async (req, res) => {
   }
 });
 
-//Update Idea
+// Update idea
 router.put('/:id', async (req, res) => {
   try {
-    const updateIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updateIdea });
+        { new: true }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorized to update this resource',
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
   }
 });
 
-//Delete Idea
+// Delete idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorized to delete this resource',
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
